feat:ability to keep track of how many times a link is visited
