+++ package.json
@@ -43,6 +43,7 @@
     "react-apexcharts": "^1.4.0",
     "react-dom": "18.2.0",
     "styled-components": "^5.3.6",
+    "turl": "^1.1.2",
     "typescript": "5.0.3",
     "ws": "7.0.0"
   },
